refactor(multi_agent_simulation.py): moved chart and video creation methods to chart.py
